+++ src/App/pages/RecipesListPage/RecipesListPage.tsx
@@ -1,18 +1,16 @@
 import { useEffect, useState } from 'react';
 
-import recipesListBg from 'assets/recipes-list-bg.jpg';
-import recipesListTitle from 'assets/recipes-list-title.png';
-
 import Loader from 'components/Loader';
-import Text from 'components/Text';
 
 // import FoodService from 'services/FoodService';
 import { getTestRecipes } from 'utils/getTestRecipes';
 import { IRecipeListItem } from 'utils/types';
 
 import Filters from './components/Filters';
+import Header from './components/Header';
 import Paginator from './components/Paginator';
 import RecipesList from './components/RecipesList';
+import Subtitle from './components/Subtitle';
 
 import styles from './RecipesListPage.module.scss';
 
@@ -62,20 +60,11 @@ const RecipesListPage = () => {
     <div className={`container ${styles['recipes-list']}`}>
       {!isLoading ? (
         <>
-          <div className={styles['recipes-list__header']}>
-            <img className={styles['recipes-list__bg']} src={recipesListBg} alt="recipes-list-bg" />
-            <img className={styles['recipes-list__title']} src={recipesListTitle} alt="recipes-list-title" />
-          </div>
-
-          <Text className={styles['recipes-list__subtitle']} tag="h2" weight="normal" view="p-20">
-            Find the perfect food and <u>drink ideas</u> for every occasion, from <u>weeknight dinners</u> to{' '}
-            <u>holiday feasts</u>.
-          </Text>
-
+          <Header />
+          <Subtitle />
           <Filters />
 
           <RecipesList recipesList={recipesList} />
-
           <Paginator page={1} pages={9} onPageSwitch={() => {}} />
         </>
       ) : (
