+++ src/App/pages/RecipesListPage/components/AlarmClockIcon/AlarmClockIcon.module.scss
@@ -1,4 +1,4 @@
-@use "../../../styles/variables" as *;
+@use "../../../../../styles/variables" as *;
 
 .color-text-primary {
   stroke: $color-text-primary;
