+++ src/App/pages/RecipesListPage/components/AlarmClockIcon/AlarmClockIcon.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import Icon, { IconProps } from '../Icon';
+import Icon, { IconProps } from 'components/icons/Icon';
 
 import styles from './AlarmClockIcon.module.scss';
