+++ src/App/pages/RecipesListPage/components/Header/Header.module.scss
@@ -1,0 +1,27 @@
+@use "../../../../../styles/variables" as *;
+
+.header {
+  position: relative;
+  top: 0;
+  left: -$container-padding;
+  width: 100vw;
+  height: 572px;
+
+  &__bg {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100vw;
+    height: 100%;
+    object-fit: cover;
+    z-index: 10;
+  }
+
+  &__title {
+    position: absolute;
+    bottom: 20px;
+    left: 50%;
+    transform: translateX(-50%);
+    z-index: 100;
+  }
+}
