+++ src/App/pages/RecipesListPage/components/Subtitle/Subtitle.module.scss
@@ -1,0 +1,6 @@
+@use "../../../../../styles/variables" as *;
+
+.subtitle {
+  margin-top: $space-h;
+  text-align: center;
+}
