+++ src/App/pages/RecipesListPage/components/Subtitle/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Subtitle';
